+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+<a name="1.1.1"></a>
+## 1.1.1 (2022-09-02)
+
+### Added
+
+- ✨ Add an option to provide json on command line
+- ✨ Add an option to provide selector on command line
+
+### Changed
+- ⬆️ Upgrade node-html-to-image
+
 <a name="1.1.0"></a>
 ## 1.1.0 (2020-03-24)